+++ compshop/src/main/java/me/vinuvicho/compshop/repository/OrderRepo.java
@@ -8,10 +8,10 @@ import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Repository
-public interface OrderRepo extends JpaRepository<Order, Long> {
+public interface OrderRepo extends JpaRepository<Order, Integer> {
 
     @Modifying
     @Transactional
     @Query("update Computer set stock = stock - 1 where computerId = ?1")
-    void decreaseComputerStock(Long computerId);
+    void decreaseComputerStock(Integer computerId);
 }
